Release v5.6.0

+++ include/keepkey/firmware/coins.h
@@ -41,7 +41,7 @@ extern const CoinTable coins;
 
 const CoinType *coinByShortcut(const char *shortcut);
 const CoinType *coinByName(const char *name);
-const CoinType *coinByAddressType(uint8_t address_type);
+const CoinType *coinByAddressType(uint32_t address_type);
 void coin_amnt_to_str(const CoinType *coin, uint64_t amnt, char *buf, int len);
 bool bip44_node_to_string(const CoinType *coin, char *node_str, uint32_t *address_n,
                          size_t address_n_count);
